interverse/intersight/tests/pyproject.toml
[project]
name = "intersight-tests"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pytest>=8.0", "pyyaml>=6.0"]

[tool.pytest.ini_options]
testpaths = ["structural"]
